backend/pyproject.toml
[project]
name = "conversational-learning-backend"
version = "0.1.0"
description = "AI-powered conversational learning assistant for visually impaired students"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "openai-whisper>=20231117",
    "pillow>=10.3.0",
    "pydantic>=2.7.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "soundfile>=0.12.1",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
